tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-manual.html

<!DOCTYPE html>
<title>Element#requestFullscreen()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    var div = document.querySelector("div");

    document.addEventListener("fullscreenchange", t.step_func(function(event)
    {
        assert_equals(event.target, document, "event.target");
        assert_false(event.bubbles, "event.bubbles");
        assert_false(event.cancelable, "event.cancelable");
        t.done();
    }));

    trusted_request(t, div);
});
</script>
